requirements/prod.txt
django==5.2b1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
gunicorn==23.0.0
pycountry==24.6.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
